[onert] Rename loader file and namespace
This commit renames loader file and namespace.

ONE-DCO-1.0-Signed-off-by: Hyeongseok Oh <hseok82.oh@samsung.com>
