+++ runtime/onert/core/include/loader/TrainInfoLoader.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Samsung Electronics Co., Ltd. All Rights Reserved
+ * Copyright (c) 2024 Samsung Electronics Co., Ltd. All Rights Reserved
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,24 @@
  * limitations under the License.
  */
 
-#ifndef __CIRCLE_CIRCLE_LOADER_H__
-#define __CIRCLE_CIRCLE_LOADER_H__
+#ifndef __ONERT_LOADER_CIRCLE_TRAININFO_LOADER_H__
+#define __ONERT_LOADER_CIRCLE_TRAININFO_LOADER_H__
 
-#include "ir/Graph.h"
+#include "ir/train/TrainingInfo.h"
 
 #include <memory>
 
 namespace onert
 {
-namespace circle_loader
+namespace loader
 {
-std::unique_ptr<ir::Model> loadModel(const std::string &filename);
-std::unique_ptr<ir::Model> loadModel(uint8_t *buffer, size_t size);
-} // namespace circle_loader
+
+// TODO change this line to use inline variable after C++17
+extern const char *const TRAININFO_METADATA_NAME;
+
+std::unique_ptr<ir::train::TrainingInfo> loadTrainingInfo(const uint8_t *buffer, const size_t size);
+
+} // namespace loader
 } // namespace onert
 
-#endif // __CIRCLE_CIRCLE_LOADER_H__
+#endif // __ONERT_LOADER_CIRCLE_TRAININFO_LOADER_H__
